refactored the create, edit, list, register and show to work with one common directory
