+++ resumator-ui/src/js/components/public/common/create.js
@@ -5,6 +5,7 @@ import NewForm from '../../shared/form/employee';
 
 import create from '../../../actions/employees/create';
 import createChange from '../../../actions/employees/createChange';
+import editCancel from '../../../actions/employees/editCancel';
 
 function mapStateToProps(state) {
   const create = state.employees.create.toJS();
@@ -20,34 +21,26 @@ function mapDispatchToProps(dispatch) {
   return {
     createEmployee: () => dispatch(create()),
     changeEmployee: (k, v) => dispatch(createChange(k, v)),
+    handleCancel: () => {
+      const really = confirm('Are you sure you want to cancel? This will discard all changes');
+      if (really) {
+        dispatch(editCancel());
+      }
+    },
   };
 }
 
-class Register extends React.Component {
-  componentWillReceiveProps(props) {
-    if (!props.session.name) {
-      return;
-    }
-
-    this.props.changeEmployee('name', props.session.name);
-    this.props.changeEmployee('surname', props.session.surname);
-    this.props.changeEmployee('email', props.session.email);
-  }
-
+class Create extends React.Component {
   render() {
-    let register;
-    if (this.props.session.name) {
-      register = true;
-    }
-
     return (
       <div>
         <NewForm
           ref="employeeForm"
-          type="EMPLOYEE"
-          register={register}
-          values={this.props.create.item}
+          type={this.props.params.type.toUpperCase()}
+          values={{}}
+          sessionValues={this.props.session}
           handleSubmit={this.props.createEmployee}
+          handleCancel={this.props.handleCancel}
           handleChange={this.props.changeEmployee.bind(this)}
           hasFailed={this.props.create.hasFailed}
           errors={this.props.create.errors}
@@ -57,4 +50,4 @@ class Register extends React.Component {
   }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Register);
+export default connect(mapStateToProps, mapDispatchToProps)(Create);
