+++ resumator-ui/src/js/components/public/common/list.js
@@ -26,7 +26,7 @@ function mapStateToProps(state) {
 
 function mapDispatchToProps(dispatch) {
   return {
-    fetchListData: () => dispatch(listAction('EMPLOYEE')),
+    fetchListData: (type) => dispatch(listAction(type)),
     removeListEntry: (type, email) => dispatch(removeAction(type, email)),
     navigateToEdit: (email) => dispatch(pushPath(`/employees/${email}/edit`, {})),
     navigateToNew: () => dispatch(pushPath('/employees/new', {})),
@@ -35,9 +35,32 @@ function mapDispatchToProps(dispatch) {
 }
 
 class List extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.type = props.params.type;
+  }
+
+  getType(props) {
+    let type = (props || this.props).params.type.toUpperCase();
+
+    type = type.substr(0, type.length - 1);
+
+    return type;
+  }
 
   componentWillMount() {
-    this.props.fetchListData();
+    this.props.fetchListData(this.getType());
+  }
+
+  componentWillReceiveProps(props) {
+    // Only make a request when the type has actually changed, to prevent an
+    // infinite loop.
+    if (props.params.type !== this.type) {
+      this.type = props.params.type;
+
+      this.props.fetchListData(this.getType(props));
+    }
   }
 
   handleRowButtonClick(email, event) {
@@ -69,7 +92,7 @@ class List extends React.Component {
       return;
     }
 
-    this.props.removeListEntry('EMPLOYEE', email);
+    this.props.removeListEntry(this.getType(), email);
   }
 
   renderTable() {
