+++ resumator-ui/src/js/components/public/common/show.js
